# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Plexus API
  description: Enterprise project management API
  version: 1.0.0
  contact:
    name: Plexus
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.plexus.app/api/v1
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Issue:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [urgent, high, medium, low, no_priority]
        status_id:
          type: string
          format: uuid
        type_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
          nullable: true
        reporter_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        sprint_id:
          type: string
          format: uuid
          nullable: true
        story_points:
          type: number
          nullable: true
        due_date:
          type: string
          format: date-time
          nullable: true
        position:
          type: number
        custom_fields:
          type: object
          additionalProperties:
            type: string
            nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, number, title, priority, status_id, type_id, reporter_id, position, created_at, updated_at]
...
  parameters:
    orgSlug:
      name: orgSlug
      in: path
      required: true
      schema:
        type: string
    projectKey:
      name: projectKey
      in: path
      required: true
      schema:
        type: string
    issueNumber:
      name: issueNumber
      in: path
      required: true
      schema:
        type: integer
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /orgs/{orgSlug}/projects/{projectKey}/issues:
    get:
      tags: [Issues]
      operationId: listIssues
      parameters:
        - name: orgSlug
          in: path
          required: true
          schema:
            type: string
        - name: projectKey
          in: path
          required: true
          schema:
            type: string
        - name: status_id
          in: query
          schema:
            type: string
        - name: assignee_id
          in: query
          schema:
            type: string
        - name: sprint_id
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of issues
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Issue'
    post:
      tags: [Issues]
      operationId: createIssue
      parameters:
        - name: orgSlug
          in: path
          required: true
          schema:
            type: string
        - name: projectKey
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                type_id:
                  type: string
                  format: uuid
                status_id:
                  type: string
                  format: uuid
                priority:
                  type: string
                  enum: [urgent, high, medium, low, no_priority]
                description:
                  type: string
                assignee_id:
                  type: string
                  format: uuid
                sprint_id:
                  type: string
                  format: uuid
                story_points:
                  type: number
                due_date:
                  type: string
                  format: date-time
                label_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
              required: [title, type_id]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  number:
                    type: integer
...
  /orgs/{orgSlug}/projects/{projectKey}/issues/{issueNumber}:
    get:
      tags: [Issues]
      operationId: getIssue
      parameters:
        - $ref: '#/components/parameters/orgSlug'
        - $ref: '#/components/parameters/projectKey'
        - $ref: '#/components/parameters/issueNumber'
      responses:
        '200':
          description: Issue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
    patch:
      tags: [Issues]
      operationId: updateIssue
      parameters:
        - $ref: '#/components/parameters/orgSlug'
        - $ref: '#/components/parameters/projectKey'
        - $ref: '#/components/parameters/issueNumber'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status_id:
                  type: string
                  format: uuid
                type_id:
                  type: string
                  format: uuid
                priority:
                  type: string
                assignee_id:
                  type: string
                  format: uuid
                sprint_id:
                  type: string
                  format: uuid
                story_points:
                  type: number
                due_date:
                  type: string
                  format: date-time
                custom_fields:
                  type: object
                  additionalProperties:
                    type: string
                    nullable: true
      responses:
        '204':
          description: Updated
    delete:
      tags: [Issues]
      operationId: deleteIssue
      parameters:
        - $ref: '#/components/parameters/orgSlug'
        - $ref: '#/components/parameters/projectKey'
        - $ref: '#/components/parameters/issueNumber'
      responses:
        '204':
          description: Deleted
...
  /orgs/{orgSlug}/projects/{projectKey}/reports/summary:
    get:
      tags: [Reports]
      operationId: projectReportSummary
      parameters:
        - $ref: '#/components/parameters/orgSlug'
        - $ref: '#/components/parameters/projectKey'
      responses:
        '200':
          description: Issue counts and story points by status category
  /orgs/{orgSlug}/projects/{projectKey}/issues/bulk:
    post:
      tags: [Issues]
      operationId: bulkUpdateIssues
      parameters:
        - $ref: '#/components/parameters/orgSlug'
        - $ref: '#/components/parameters/projectKey'
      responses:
        '200':
          description: Bulk update result
  /orgs/{orgSlug}/projects/{projectKey}/issues/import:
    post:
      tags: [Issues]
      operationId: importIssuesCSV
      parameters:
        - $ref: '#/components/parameters/orgSlug'
        - $ref: '#/components/parameters/projectKey'
      responses:
        '201':
          description: CSV import result
...